suppress stdout when running siege
Siege has a progress spinner that for an unknown reason causes
its process to hang forever in certain shells (zsh at the very least).

Luckily, siege sends most of its output to stderr, but the spinner goes
to stdout. We can safely suppress output to stdout and be okay :)

fixes hhvm/oss-performance#11