Add Memcached support for the MediaWiki target
This diff adds support for Memcached. If a target supports Memcached,
oss-performance will start it on 127.0.0.1:11888. The only target supporting
Memcached right now is MediaWiki. For other targets to support Memcached, they
must override PerfTarget::supportsMemcached().
This diff also sets $wgUseDatabaseMessages to false in the configuration file
if Memcached is disabled, as recommended in the MediaWiki manual.

For MediaWiki with Memcached enabled, observing a ~6% increase in RPS, a
decrease in system idle time, as well as ~7 percentage points more cycles being
spent in HHVM.

Command line options added:

--memcached: path to memcached binary
--delay-memcached-startup: number of seconds to delay Memcached startup
--memcached-port: TCP port to listen on (default is 11888)
--no-memcached: don't use memcached (even if target supports it)
--memcached-threads: number of memcached threads to spawn (default is 4 for
machines with small number of cores, 32 otherwise)