Mark reference arguments at the call site
HHVM 3.24 requires the reference arguments to be marked at the call site, following which the current oss-performance code throws errors when
"hh_client" is run or the workload is run using HHVM runtime. This patch fixes the issue by marking the reference arguments at the call time.

However, this patch also mandates the HHVM version to be at least 3.24 and makes it incompatible with the earlier versions as the change is
not supported in earlier versions of HHVM

Fixes https://github.com/hhvm/oss-performance/issues/89