Check for invalid arguments passed to perf.php
oss-performance does not give any warnings when passing
an incorrect argument. When a wrong argument is found, the "getopt" Hack
function that parses arguments stops silently and the rest of the arguments
are simply discarded without notice.

This change fixes this verbosity issue by going through all the parameters
passed to perf.php and checking if they are valid. If they are not, a
warning is displayed saying which parameter is wrong and the "help" message
is printed.